# docs/swagger.yaml
definitions:
...
  class.AddClassRequest:
    properties:
      credit:
        description: 学分
        type: number
      day:
        description: 星期几
        type: integer
      dur_class:
        description: 第几节 '形如 "1-3","1-1"'
        type: string
      name:
        description: 课程名称
        type: string
      semester:
        description: 学期
        type: string
      teacher:
        description: 教师
        type: string
      weeks:
        description: 哪些周
        type: integer
      where:
        description: 地点
        type: string
      year:
        description: 学年
        type: string
    type: object
  class.Class:
    properties:
      info:
        items:
          $ref: '#/definitions/class.ClassInfo'
        type: array
      thisweek:
        type: boolean
    type: object
  class.ClassInfo:
    properties:
      class_when:
        description: 上课是第几节（如1-2,3-4）
        type: string
      classname:
        description: 课程名称
        type: string
      credit:
        description: 学分
        type: number
      day:
        description: 星期几
        type: integer
      id:
        description: 集合了课程信息的字符串，便于标识（课程ID）
        type: string
      semester:
        description: 学期
        type: string
      teacher:
        description: 任课教师
        type: string
      week_duration:
        description: 上课的周数
        type: string
      weeks:
        description: 哪些周
        type: integer
      where:
        description: 上课地点
        type: string
      year:
        description: 学年
        type: string
    type: object
  class.DeleteClassRequest:
    properties:
      id:
        description: 要被删的课程id
        type: string
      semester:
        description: 学期 "1"代表第一学期，"2"代表第二学期，"3"代表第三学期
        type: string
      year:
        description: 学年  "2024" -> 代表"2024-2025学年"
        type: string
    type: object
  class.GetClassListRequest:
    properties:
      semester:
        type: string
      week:
        type: integer
      year:
        type: string
    type: object
  class.GetClassListResp:
    properties:
      classes:
        items:
          $ref: '#/definitions/class.Class'
        type: array
    type: object
  class.GetRecycleBinClassInfosResp:
    properties:
      classInfos:
        items:
          $ref: '#/definitions/class.ClassInfo'
        type: array
    type: object
  class.RecoverClassRequest:
    properties:
      classId:
        description: 课程的ID（唯一标识） 更新后这个可能会换，所以响应的时候会把新的ID返回
        type: string
      semester:
        description: 学期 "1"代表第一学期，"2"代表第二学期，"3"代表第三学期
        type: string
      year:
        description: 学年  "2024" 代表"2024-2025学年"
        type: string
    type: object
  class.SearchClassResp:
    properties:
      classInfos:
        items:
          $ref: '#/definitions/class.ClassInfo'
        type: array
    type: object
  class.SearchRequest:
    properties:
      searchKeyWords:
        description: 搜索关键词,匹配的是课程名称和教师姓名
        type: string
      semester:
        type: string
      year:
        type: string
    type: object
  class.UpdateClassRequest:
    properties:
      classId:
        description: 课程的ID（唯一标识） 更新后这个可能会换，所以响应的时候会把新的ID返回
        type: string
      credit:
        description: 学分
        type: number
      day:
        description: 星期几
        type: integer
      dur_class:
        description: 第几节 '形如 "1-3","1-1"'
        type: string
      name:
        description: 课程名称
        type: string
      semester:
        description: 学期
        type: string
      teacher:
        description: 教师
        type: string
      weeks:
        description: 哪些周
        type: integer
      where:
        description: 地点
        type: string
      year:
        description: 学年
        type: string
    type: object
  department.DelDepartmentRequest:
    properties:
      id:
        type: integer
    type: object
...
  department.SaveDepartmentRequest:
    properties:
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      place:
        type: string
      time:
        type: string
    type: object
  elecprice.CheckResponse:
    properties:
      price:
        allOf:
        - $ref: '#/definitions/elecprice.Price'
        description: 电费
    type: object
  elecprice.Price:
    properties:
      air_remain_money:
        type: string
      air_yesterday_use_money:
        type: string
      air_yesterday_use_value:
        type: string
      lighting_remain_money:
        type: string
      lighting_yesterday_use_money:
        type: string
      lighting_yesterday_use_value:
        type: string
    type: object
  elecprice.SetStandardRequest:
    properties:
      area:
        description: 区域
        type: string
      building:
        description: 建筑
        type: string
      money:
        description: 金额
        type: integer
      room:
        description: 房间号
        type: string
    type: object
...
  web.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /class/add:
    post:
      consumes:
      - application/json
      description: 添加新的课表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 课表信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/class.AddClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功添加课表
          schema:
            $ref: '#/definitions/web.Response'
      summary: 添加课表
      tags:
      - 课表
  /class/delete:
    delete:
      consumes:
      - application/json
      description: 根据课表ID删除课表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除课表请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/class.DeleteClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除课表
          schema:
            $ref: '#/definitions/web.Response'
      summary: 删除课表
      tags:
      - 课表
  /class/get:
    get:
      description: 根据学期、学年等条件获取课表
      parameters:
      - description: 学年
        in: query
        name: year
        required: true
        type: string
      - description: 学期
        in: query
        name: semester
        required: true
        type: string
      - description: 当前周
        in: query
        name: week
        required: true
        type: integer
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取课表请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class.GetClassListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回课表
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/class.GetClassListResp'
              type: object
      summary: 获取课表
      tags:
      - 课表
  /class/getRecycle:
    get:
      description: 获取已删除但未彻底清除的课表信息
      parameters:
      - description: 学年
        in: query
        name: year
        required: true
        type: string
      - description: 学期
        in: query
        name: semester
        required: true
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取回收站课表信息
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/class.GetRecycleBinClassInfosResp'
              type: object
      summary: 获取回收站课表信息
      tags:
      - 课表
  /class/recover:
    put:
      consumes:
      - application/json
      description: 从回收站恢复课表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 恢复课表请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/class.RecoverClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功恢复课表
          schema:
            $ref: '#/definitions/web.Response'
      summary: 恢复课表
      tags:
      - 课表
  /class/search:
    get:
      description: 根据关键词搜索课程
      parameters:
      - description: 学年
        in: query
        name: year
        required: true
        type: string
      - description: 学期
        in: query
        name: semester
        required: true
        type: string
      - description: 搜索关键词
        in: query
        name: searchKeyWords
        required: true
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 搜索课程请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/class.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功搜索到课程
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/class.SearchClassResp'
              type: object
      summary: 搜索课程
      tags:
      - 课表
  /class/update:
    put:
      consumes:
      - application/json
      description: 根据课表ID更新课表信息
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新课表请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/class.UpdateClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新课表
          schema:
            $ref: '#/definitions/web.Response'
      summary: 更新课表信息
      tags:
      - 课表
  /department/delDepartment:
    delete:
      consumes:
      - application/json
      description: 删除部门信息
      parameters:
      - description: 删除部门信息请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/department.DelDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
      summary: 删除部门信息
      tags:
      - 部门
...
  /department/saveDepartment:
    post:
      consumes:
      - application/json
      description: 保存部门信息
      parameters:
      - description: 保存部门信息请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/department.SaveDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
      summary: 保存部门信息
      tags:
      - 部门
  /elecprice/check:
    get:
      description: 根据区域、楼栋和房间号查询电费信息
      parameters:
      - description: 区域，例如 '东区学生宿舍'
        in: query
        name: area
        required: true
        type: string
      - description: 楼栋，例如 '1号楼'
        in: query
        name: building
        required: true
        type: string
      - description: 房间号，例如 '101'
        in: query
        name: room
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回电费信息
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/elecprice.CheckResponse'
              type: object
        "500":
          description: 系统异常
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 查询电费
      tags:
      - 电费
  /elecprice/setStandard:
    post:
      consumes:
      - application/json
      description: 根据区域、楼栋和房间号设置电费提醒的金额标准
      parameters:
      - description: 设置电费提醒请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/elecprice.SetStandardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 设置成功的返回信息
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                msg:
                  type: string
              type: object
        "500":
          description: 系统异常
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 设置电费提醒标准
      tags:
      - 电费
...
swagger: "2.0"